Add fuse_dq_q_pass in exir/passes and also add it to HTP backend (#2295)
Summary:

There are passes such as these https://fburl.com/code/vs6n4vcv that remove noops from the graph. The problem is that after this pass runs it still leaves in the dq->q nodes in the graph which then potentially get delegated to the backend causing perf regressions.

This pass will remove the dq->q ops if their qparams are of the same value. If not it won't touch them.

Differential Revision: D54543323